# Optimize turbulence noise