--- a/Dev/Cpp/Effekseer/Effekseer/SIMD/Effekseer.SIMD4f.h
+++ b/Dev/Cpp/Effekseer/Effekseer/SIMD/Effekseer.SIMD4f.h
@@ -3,14 +3,14 @@
 #define __EFFEKSEER_SIMD4F_H__
 
 #include <cstdint>
-#include <cmath>
+#include "Effekseer.SIMDType.h"
 
-//#if defined(__ARM_NEON__) || defined(__ARM_NEON)
-//#include "Effekseer.SIMD4f_NEON.h"
-//#elif (defined(_M_AMD64) || defined(_M_X64)) || (defined(_M_IX86_FP) && _M_IX86_FP >= 2) || defined(__SSE__)
-//#include "Effekseer.SIMD4f_SSE.h"
-//#else
+#if defined(EFK_SIMD_NEON)
+#include "Effekseer.SIMD4f_NEON.h"
+#elif defined(EFK_SIMD_SSE2)
+#include "Effekseer.SIMD4f_SSE.h"
+#else
 #include "Effekseer.SIMD4f_Gen.h"
-//#endif
+#endif
 
 #endif // __EFFEKSEER_SIMD4F_H__