--- a/Dev/Cpp/Effekseer/Effekseer/SIMD/Effekseer.SIMD4f_Gen.h
+++ b/Dev/Cpp/Effekseer/Effekseer/SIMD/Effekseer.SIMD4f_Gen.h
@@ -2,11 +2,11 @@
 #ifndef __EFFEKSEER_SIMD4F_GEN_H__
 #define __EFFEKSEER_SIMD4F_GEN_H__
 
-#if defined(__ARM_NEON__) || defined(__ARM_NEON)
-// not arm
-#elif (defined(_M_AMD64) || defined(_M_X64)) || (defined(_M_IX86_FP) && _M_IX86_FP >= 2) || defined(__SSE__)
-// not x86
-#else
+//#if defined(__ARM_NEON__) || defined(__ARM_NEON)
+//// not arm
+//#elif (defined(_M_AMD64) || defined(_M_X64)) || (defined(_M_IX86_FP) && _M_IX86_FP >= 2) || defined(__SSE__)
+//// not x86
+//#else
 
 #include <stdint.h>
 #include <algorithm>
@@ -24,36 +24,45 @@ inline float Rsqrt(float x)
 	return 1.0f / std::sqrt(x);
 }
 
+struct SIMD4i;
+
 /**
 	@brief	simd class for generic
 */
 struct alignas(16) SIMD4f
 {
 	union {
-		float f[4];
-		uint32_t i[4];
+		float vf[4];
+		int32_t vi[4];
+		uint32_t vu[4];
 	};
 
 	SIMD4f() = default;
 	SIMD4f(const SIMD4f& rhs) = default;
-	SIMD4f(float x, float y, float z, float w) { f[0] = x; f[1] = y; f[2] = z; f[3] = w; }
-	SIMD4f(float i) { f[0] = i; f[1] = i; f[2] = i; f[3] = i; }
+	SIMD4f(float x, float y, float z, float w) { vf[0] = x; vf[1] = y; vf[2] = z; vf[3] = w; }
+	SIMD4f(float i) { vf[0] = i; vf[1] = i; vf[2] = i; vf[3] = i; }
 
-	float GetX() const { return f[0]; }
-	float GetY() const { return f[1]; }
-	float GetZ() const { return f[2]; }
-	float GetW() const { return f[3]; }
+	float GetX() const { return vf[0]; }
+	float GetY() const { return vf[1]; }
+	float GetZ() const { return vf[2]; }
+	float GetW() const { return vf[3]; }
 
-	void SetX(float o) { f[0] = o; }
-	void SetY(float o) { f[1] = o; }
-	void SetZ(float o) { f[2] = o; }
-	void SetW(float o) { f[3] = o; }
+	void SetX(float o) { vf[0] = o; }
+	void SetY(float o) { vf[1] = o; }
+	void SetZ(float o) { vf[2] = o; }
+	void SetW(float o) { vf[3] = o; }
+
+	template <size_t LANE>
+	SIMD4f Dup() { return SIMD4f(vf[LANE], vf[LANE], vf[LANE], vf[LANE]); }
+
+	SIMD4i Convert4i() const;
+	SIMD4i Cast4i() const;
 
 	SIMD4f& operator+=(const SIMD4f& rhs)
 	{
 		for (size_t i = 0; i < 4; i++)
 		{
-			f[i] += rhs.f[i];
+			vf[i] += rhs.vf[i];
 		}
 		return *this;
 	}
@@ -62,7 +71,7 @@ struct alignas(16) SIMD4f
 	{
 		for (size_t i = 0; i < 4; i++)
 		{
-			f[i] -= rhs.f[i];
+			vf[i] -= rhs.vf[i];
 		}
 		return *this;
 	}
@@ -71,7 +80,7 @@ struct alignas(16) SIMD4f
 	{
 		for (size_t i = 0; i < 4; i++)
 		{
-			f[i] *= rhs.f[i];
+			vf[i] *= rhs.vf[i];
 		}
 		return *this;
 	}
@@ -80,7 +89,7 @@ struct alignas(16) SIMD4f
 	{
 		for (size_t i = 0; i < 4; i++)
 		{
-			f[i] *= rhs;
+			vf[i] *= rhs;
 		}
 		return *this;
 	}
@@ -89,7 +98,7 @@ struct alignas(16) SIMD4f
 	{
 		for (size_t i = 0; i < 4; i++)
 		{
-			f[i] /= rhs.f[i];
+			vf[i] /= rhs.vf[i];
 		}
 		return *this;
 	}
@@ -98,7 +107,7 @@ struct alignas(16) SIMD4f
 	{
 		for (size_t i = 0; i < 4; i++)
 		{
-			f[i] /= rhs;
+			vf[i] /= rhs;
 		}
 		return *this;
 	}
@@ -118,6 +127,8 @@ struct alignas(16) SIMD4f
 	static SIMD4f Abs(const SIMD4f& in);
 	static SIMD4f Min(const SIMD4f& lhs, const SIMD4f& rhs);
 	static SIMD4f Max(const SIMD4f& lhs, const SIMD4f& rhs);
+	static SIMD4f Floor(const SIMD4f& in);
+	static SIMD4f Ceil(const SIMD4f& in);
 	static SIMD4f MulAdd(const SIMD4f& a, const SIMD4f& b, const SIMD4f& c);
 	static SIMD4f MulSub(const SIMD4f& a, const SIMD4f& b, const SIMD4f& c);
 
@@ -136,6 +147,7 @@ struct alignas(16) SIMD4f
 	template <uint32_t X, uint32_t Y, uint32_t Z, uint32_t W>
 	static SIMD4f Mask();
 	static uint32_t MoveMask(const SIMD4f& in);
+	static SIMD4f Select(const SIMD4f& mask, const SIMD4f& sel1, const SIMD4f& sel2);
 	static SIMD4f Equal(const SIMD4f& lhs, const SIMD4f& rhs);
 	static SIMD4f NotEqual(const SIMD4f& lhs, const SIMD4f& rhs);
 	static SIMD4f LessThan(const SIMD4f& lhs, const SIMD4f& rhs);
@@ -147,12 +159,29 @@ struct alignas(16) SIMD4f
 	static void Transpose(SIMD4f& s0, SIMD4f& s1, SIMD4f& s2, SIMD4f& s3);
 };
 
+} // namespace Effekseer
+
+#include "Effekseer.SIMD4i.h"
+
+namespace Effekseer
+{
+
+inline SIMD4i SIMD4f::Convert4i() const
+{
+	return SIMD4i((int32_t)vf[0], (int32_t)vf[1], (int32_t)vf[2], (int32_t)vf[3]);
+}
+
+inline SIMD4i SIMD4f::Cast4i() const
+{
+	return SIMD4i(vu[0], vu[1], vu[2], vu[3]);
+}
+
 inline SIMD4f operator+(const SIMD4f& lhs, const SIMD4f& rhs)
 {
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.f[i] = lhs.f[i] + rhs.f[i];
+		ret.vf[i] = lhs.vf[i] + rhs.vf[i];
 	}
 	return ret;
 }
@@ -162,7 +191,7 @@ inline SIMD4f operator-(const SIMD4f& lhs, const SIMD4f& rhs)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.f[i] = lhs.f[i] - rhs.f[i];
+		ret.vf[i] = lhs.vf[i] - rhs.vf[i];
 	}
 	return ret;
 }
@@ -172,7 +201,7 @@ inline SIMD4f operator*(const SIMD4f& lhs, const SIMD4f& rhs)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.f[i] = lhs.f[i] * rhs.f[i];
+		ret.vf[i] = lhs.vf[i] * rhs.vf[i];
 	}
 	return ret;
 }
@@ -182,7 +211,7 @@ inline SIMD4f operator*(const SIMD4f& lhs, float rhs)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.f[i] = lhs.f[i] * rhs;
+		ret.vf[i] = lhs.vf[i] * rhs;
 	}
 	return ret;
 }
@@ -192,7 +221,7 @@ inline SIMD4f operator/(const SIMD4f& lhs, const SIMD4f& rhs)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.f[i] = lhs.f[i] / rhs.f[i];
+		ret.vf[i] = lhs.vf[i] / rhs.vf[i];
 	}
 	return ret;
 }
@@ -202,7 +231,7 @@ inline SIMD4f operator/(const SIMD4f& lhs, float rhs)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.f[i] = lhs.f[i] / rhs;
+		ret.vf[i] = lhs.vf[i] / rhs;
 	}
 	return ret;
 }
@@ -212,7 +241,7 @@ inline SIMD4f operator&(const SIMD4f& lhs, const SIMD4f& rhs)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.i[i] = lhs.i[i] & rhs.i[i];
+		ret.vu[i] = lhs.vu[i] & rhs.vu[i];
 	}
 	return ret;
 }
@@ -222,7 +251,17 @@ inline SIMD4f operator|(const SIMD4f& lhs, const SIMD4f& rhs)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.i[i] = lhs.i[i] | rhs.i[i];
+		ret.vu[i] = lhs.vu[i] | rhs.vu[i];
+	}
+	return ret;
+}
+
+inline SIMD4f operator^(const SIMD4f& lhs, const SIMD4f& rhs)
+{
+	SIMD4f ret;
+	for (size_t i = 0; i < 4; i++)
+	{
+		ret.vu[i] = lhs.vu[i] ^ rhs.vu[i];
 	}
 	return ret;
 }
@@ -232,7 +271,7 @@ inline bool operator==(const SIMD4f& lhs, const SIMD4f& rhs)
 	bool ret = true;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret &= lhs.f[i] == rhs.f[i];
+		ret &= lhs.vf[i] == rhs.vf[i];
 	}
 	return ret;
 }
@@ -242,7 +281,7 @@ inline bool operator!=(const SIMD4f& lhs, const SIMD4f& rhs)
 	bool ret = true;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret &= lhs.f[i] == rhs.f[i];
+		ret &= lhs.vf[i] == rhs.vf[i];
 	}
 	return !ret;
 }
@@ -250,90 +289,90 @@ inline bool operator!=(const SIMD4f& lhs, const SIMD4f& rhs)
 inline SIMD4f SIMD4f::Load2(const void* mem)
 {
 	SIMD4f ret;
-	memcpy(ret.f, mem, sizeof(float) * 2);
+	memcpy(ret.vf, mem, sizeof(float) * 2);
 	// This code causes bugs in asmjs
-	// ret.f[0] = *((float*)mem + 0);
-	// ret.f[1] = *((float*)mem + 1);
+	// ret.vf[0] = *((float*)mem + 0);
+	// ret.vf[1] = *((float*)mem + 1);
 	return ret;
 }
 
 inline void SIMD4f::Store2(void* mem, const SIMD4f& i)
 {
-	memcpy(mem, i.f, sizeof(float) * 2);
+	memcpy(mem, i.vf, sizeof(float) * 2);
 	// This code causes bugs in asmjs
-	// *((float*)mem + 0) = i.f[0];
-	// *((float*)mem + 1) = i.f[1];
+	// *((float*)mem + 0) = i.vf[0];
+	// *((float*)mem + 1) = i.vf[1];
 }
 
 inline SIMD4f SIMD4f::Load3(const void* mem)
 {
 	SIMD4f ret;
-	memcpy(ret.f, mem, sizeof(float) * 3);
+	memcpy(ret.vf, mem, sizeof(float) * 3);
 	// This code causes bugs in asmjs
-	// ret.f[0] = *((float*)mem + 0);
-	// ret.f[1] = *((float*)mem + 1);
-	// ret.f[2] = *((float*)mem + 2);
+	// ret.vf[0] = *((float*)mem + 0);
+	// ret.vf[1] = *((float*)mem + 1);
+	// ret.vf[2] = *((float*)mem + 2);
 	return ret;
 }
 
 inline void SIMD4f::Store3(void* mem, const SIMD4f& i)
 {
-	memcpy(mem, i.f, sizeof(float) * 3);
+	memcpy(mem, i.vf, sizeof(float) * 3);
 	// This code causes bugs in asmjs
-	// *((float*)mem + 0) = i.f[0];
-	// *((float*)mem + 1) = i.f[1];
-	// *((float*)mem + 2) = i.f[2];
+	// *((float*)mem + 0) = i.vf[0];
+	// *((float*)mem + 1) = i.vf[1];
+	// *((float*)mem + 2) = i.vf[2];
 }
 
 inline SIMD4f SIMD4f::Load4(const void* mem)
 {
 	SIMD4f ret;
-	memcpy(ret.f, mem, sizeof(float) * 4);
+	memcpy(ret.vf, mem, sizeof(float) * 4);
 	// This code causes bugs in emscripten
-	// ret.f[0] = *((float*)mem + 0);
-	// ret.f[1] = *((float*)mem + 1);
-	// ret.f[2] = *((float*)mem + 2);
-	// ret.f[3] = *((float*)mem + 3);
+	// ret.vf[0] = *((float*)mem + 0);
+	// ret.vf[1] = *((float*)mem + 1);
+	// ret.vf[2] = *((float*)mem + 2);
+	// ret.vf[3] = *((float*)mem + 3);
 	return ret;
 }
 
 inline void SIMD4f::Store4(void* mem, const SIMD4f& i)
 {
-	memcpy(mem, i.f, sizeof(float) * 4);
+	memcpy(mem, i.vf, sizeof(float) * 4);
 	// This code causes bugs in asmjs
-	// *((float*)mem + 0) = i.f[0];
-	// *((float*)mem + 1) = i.f[1];
-	// *((float*)mem + 2) = i.f[2];
-	// *((float*)mem + 3) = i.f[3];
+	// *((float*)mem + 0) = i.vf[0];
+	// *((float*)mem + 1) = i.vf[1];
+	// *((float*)mem + 2) = i.vf[2];
+	// *((float*)mem + 3) = i.vf[3];
 }
 
 inline SIMD4f SIMD4f::SetZero()
 {
 	SIMD4f ret;
-	ret.f[0] = 0.0f;
-	ret.f[1] = 0.0f;
-	ret.f[2] = 0.0f;
-	ret.f[3] = 0.0f;
+	ret.vf[0] = 0.0f;
+	ret.vf[1] = 0.0f;
+	ret.vf[2] = 0.0f;
+	ret.vf[3] = 0.0f;
 	return ret;
 }
 
 inline SIMD4f SIMD4f::SetInt(int32_t x, int32_t y, int32_t z, int32_t w)
 {
 	SIMD4f ret;
-	ret.i[0] = (uint32_t)x;
-	ret.i[1] = (uint32_t)y;
-	ret.i[2] = (uint32_t)z;
-	ret.i[3] = (uint32_t)w;
+	ret.vu[0] = (uint32_t)x;
+	ret.vu[1] = (uint32_t)y;
+	ret.vu[2] = (uint32_t)z;
+	ret.vu[3] = (uint32_t)w;
 	return ret;
 }
 
 inline SIMD4f SIMD4f::SetUInt(uint32_t x, uint32_t y, uint32_t z, uint32_t w)
 {
 	SIMD4f ret;
-	ret.i[0] = (uint32_t)x;
-	ret.i[1] = (uint32_t)y;
-	ret.i[2] = (uint32_t)z;
-	ret.i[3] = (uint32_t)w;
+	ret.vu[0] = (uint32_t)x;
+	ret.vu[1] = (uint32_t)y;
+	ret.vu[2] = (uint32_t)z;
+	ret.vu[3] = (uint32_t)w;
 	return ret;
 }
 
@@ -342,7 +381,7 @@ inline SIMD4f SIMD4f::Sqrt(const SIMD4f& in)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.f[i] = std::sqrt(in.f[i]);
+		ret.vf[i] = std::sqrt(in.vf[i]);
 	}
 	return ret;
 }
@@ -352,7 +391,7 @@ inline SIMD4f SIMD4f::Rsqrt(const SIMD4f& in)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.f[i] = 1.0f / std::sqrt(in.f[i]);
+		ret.vf[i] = 1.0f / std::sqrt(in.vf[i]);
 	}
 	return ret;
 }
@@ -362,7 +401,7 @@ inline SIMD4f SIMD4f::Abs(const SIMD4f& in)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.f[i] = std::abs(in.f[i]);
+		ret.vf[i] = std::abs(in.vf[i]);
 	}
 	return ret;
 }
@@ -372,7 +411,7 @@ inline SIMD4f SIMD4f::Min(const SIMD4f& lhs, const SIMD4f& rhs)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.f[i] = std::fmin(lhs.f[i], rhs.f[i]);
+		ret.vf[i] = std::fmin(lhs.vf[i], rhs.vf[i]);
 	}
 	return ret;
 }
@@ -382,7 +421,27 @@ inline SIMD4f SIMD4f::Max(const SIMD4f& lhs, const SIMD4f& rhs)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.f[i] = std::fmax(lhs.f[i], rhs.f[i]);
+		ret.vf[i] = std::fmax(lhs.vf[i], rhs.vf[i]);
+	}
+	return ret;
+}
+
+inline SIMD4f SIMD4f::Floor(const SIMD4f& in)
+{
+	SIMD4f ret;
+	for (size_t i = 0; i < 4; i++)
+	{
+		ret.vf[i] = std::floor(in.vf[i]);
+	}
+	return ret;
+}
+
+inline SIMD4f SIMD4f::Ceil(const SIMD4f& in)
+{
+	SIMD4f ret;
+	for (size_t i = 0; i < 4; i++)
+	{
+		ret.vf[i] = std::ceil(in.vf[i]);
 	}
 	return ret;
 }
@@ -392,7 +451,7 @@ inline SIMD4f SIMD4f::MulAdd(const SIMD4f& a, const SIMD4f& b, const SIMD4f& c)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.f[i] = a.f[i] + b.f[i] * c.f[i];
+		ret.vf[i] = a.vf[i] + b.vf[i] * c.vf[i];
 }
 	return ret;
 }
@@ -402,7 +461,7 @@ inline SIMD4f SIMD4f::MulSub(const SIMD4f& a, const SIMD4f& b, const SIMD4f& c)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.f[i] = a.f[i] - b.f[i] * c.f[i];
+		ret.vf[i] = a.vf[i] - b.vf[i] * c.vf[i];
 }
 	return ret;
 }
@@ -410,7 +469,7 @@ inline SIMD4f SIMD4f::MulSub(const SIMD4f& a, const SIMD4f& b, const SIMD4f& c)
 inline SIMD4f SIMD4f::Dot3(const SIMD4f& lhs, const SIMD4f& rhs)
 {
 	SIMD4f muled = lhs * rhs;
-	return SIMD4f{muled.f[0] + muled.f[1] + muled.f[2], 0.0f, 0.0f, 0.0f};
+	return SIMD4f{muled.vf[0] + muled.vf[1] + muled.vf[2], 0.0f, 0.0f, 0.0f};
 }
 
 inline SIMD4f SIMD4f::Cross3(const SIMD4f& lhs, const SIMD4f& rhs)
@@ -423,21 +482,21 @@ template<size_t LANE>
 SIMD4f SIMD4f::MulLane(const SIMD4f& lhs, const SIMD4f& rhs)
 {
 	static_assert(LANE < 4, "LANE is must be less than 4.");
-	return lhs * rhs.f[LANE];
+	return lhs * rhs.vf[LANE];
 }
 
 template<size_t LANE>
 SIMD4f SIMD4f::MulAddLane(const SIMD4f& a, const SIMD4f& b, const SIMD4f& c)
 {
 	static_assert(LANE < 4, "LANE is must be less than 4.");
-	return a + b * c.f[LANE];
+	return a + b * c.vf[LANE];
 }
 
 template<size_t LANE>
 SIMD4f SIMD4f::MulSubLane(const SIMD4f& a, const SIMD4f& b, const SIMD4f& c)
 {
 	static_assert(LANE < 4, "LANE is must be less than 4.");
-	return a - b * c.f[LANE];
+	return a - b * c.vf[LANE];
 }
 
 template <uint32_t indexX, uint32_t indexY, uint32_t indexZ, uint32_t indexW>
@@ -447,7 +506,7 @@ SIMD4f SIMD4f::Swizzle(const SIMD4f& in)
 	static_assert(indexY < 4, "indexY is must be less than 4.");
 	static_assert(indexZ < 4, "indexZ is must be less than 4.");
 	static_assert(indexW < 4, "indexW is must be less than 4.");
-	return SIMD4f{in.f[indexX], in.f[indexY], in.f[indexZ], in.f[indexW]};
+	return SIMD4f{in.vf[indexX], in.vf[indexY], in.vf[indexZ], in.vf[indexW]};
 }
 
 
@@ -455,16 +514,26 @@ template <uint32_t X, uint32_t Y, uint32_t Z, uint32_t W>
 SIMD4f SIMD4f::Mask()
 {
 	SIMD4f ret;
-	ret.i[0] = 0xffffffff * X;
-	ret.i[1] = 0xffffffff * Y;
-	ret.i[2] = 0xffffffff * Z;
-	ret.i[3] = 0xffffffff * W;
+	ret.vu[0] = 0xffffffff * X;
+	ret.vu[1] = 0xffffffff * Y;
+	ret.vu[2] = 0xffffffff * Z;
+	ret.vu[3] = 0xffffffff * W;
 	return ret;
 }
 
 inline uint32_t SIMD4f::MoveMask(const SIMD4f& in)
 {
-	return (in.i[0] & 0x1) | (in.i[1] & 0x2) | (in.i[2] & 0x4) | (in.i[3] & 0x8);
+	return (in.vu[0] & 0x1) | (in.vu[1] & 0x2) | (in.vu[2] & 0x4) | (in.vu[3] & 0x8);
+}
+
+inline SIMD4f SIMD4f::Select(const SIMD4f& mask, const SIMD4f& sel1, const SIMD4f& sel2)
+{
+	SIMD4f ret;
+	for (size_t i = 0; i < 4; i++)
+	{
+		ret.vu[i] = (mask.vu[i] & sel1.vu[i]) | (~mask.vu[i] & sel2.vu[i]);
+	}
+	return ret;
 }
 
 inline SIMD4f SIMD4f::Equal(const SIMD4f& lhs, const SIMD4f& rhs)
@@ -472,7 +541,7 @@ inline SIMD4f SIMD4f::Equal(const SIMD4f& lhs, const SIMD4f& rhs)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.i[i] = (lhs.f[i] == rhs.f[i]) ? 0xffffffff : 0;
+		ret.vu[i] = (lhs.vf[i] == rhs.vf[i]) ? 0xffffffff : 0;
 	}
 	return ret;
 }
@@ -482,7 +551,7 @@ inline SIMD4f SIMD4f::NotEqual(const SIMD4f& lhs, const SIMD4f& rhs)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.i[i] = (lhs.f[i] != rhs.f[i]) ? 0xffffffff : 0;
+		ret.vu[i] = (lhs.vf[i] != rhs.vf[i]) ? 0xffffffff : 0;
 	}
 	return ret;
 }
@@ -492,7 +561,7 @@ inline SIMD4f SIMD4f::LessThan(const SIMD4f& lhs, const SIMD4f& rhs)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.i[i] = (lhs.f[i] < rhs.f[i]) ? 0xffffffff : 0;
+		ret.vu[i] = (lhs.vf[i] < rhs.vf[i]) ? 0xffffffff : 0;
 	}
 	return ret;
 }
@@ -502,7 +571,7 @@ inline SIMD4f SIMD4f::LessEqual(const SIMD4f& lhs, const SIMD4f& rhs)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.i[i] = (lhs.f[i] <= rhs.f[i]) ? 0xffffffff : 0;
+		ret.vu[i] = (lhs.vf[i] <= rhs.vf[i]) ? 0xffffffff : 0;
 	}
 	return ret;
 }
@@ -512,7 +581,7 @@ inline SIMD4f SIMD4f::GreaterThan(const SIMD4f& lhs, const SIMD4f& rhs)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.i[i] = (lhs.f[i] > rhs.f[i]) ? 0xffffffff : 0;
+		ret.vu[i] = (lhs.vf[i] > rhs.vf[i]) ? 0xffffffff : 0;
 	}
 	return ret;
 }
@@ -522,7 +591,7 @@ inline SIMD4f SIMD4f::GreaterEqual(const SIMD4f& lhs, const SIMD4f& rhs)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.i[i] = (lhs.f[i] >= rhs.f[i]) ? 0xffffffff : 0;
+		ret.vu[i] = (lhs.vf[i] >= rhs.vf[i]) ? 0xffffffff : 0;
 	}
 	return ret;
 }
@@ -532,7 +601,7 @@ inline SIMD4f SIMD4f::NearEqual(const SIMD4f& lhs, const SIMD4f& rhs, float epsi
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.i[i] = (std::abs(lhs.f[i] - rhs.f[i]) <= epsilon) ? 0xffffffff : 0;
+		ret.vu[i] = (std::abs(lhs.vf[i] - rhs.vf[i]) <= epsilon) ? 0xffffffff : 0;
 	}
 	return ret;
 }
@@ -542,23 +611,23 @@ inline SIMD4f SIMD4f::IsZero(const SIMD4f& in, float epsilon)
 	SIMD4f ret;
 	for (size_t i = 0; i < 4; i++)
 	{
-		ret.i[i] = (std::abs(in.f[i]) <= epsilon) ? 0xffffffff : 0;
+		ret.vu[i] = (std::abs(in.vf[i]) <= epsilon) ? 0xffffffff : 0;
 	}
 	return ret;
 }
 
 inline void SIMD4f::Transpose(SIMD4f& s0, SIMD4f& s1, SIMD4f& s2, SIMD4f& s3)
 {
-	std::swap(s0.f[1], s1.f[0]);
-	std::swap(s0.f[2], s2.f[0]);
-	std::swap(s0.f[3], s3.f[0]);
-	std::swap(s1.f[2], s2.f[1]);
-	std::swap(s2.f[3], s3.f[2]);
-	std::swap(s1.f[3], s3.f[1]);
+	std::swap(s0.vf[1], s1.vf[0]);
+	std::swap(s0.vf[2], s2.vf[0]);
+	std::swap(s0.vf[3], s3.vf[0]);
+	std::swap(s1.vf[2], s2.vf[1]);
+	std::swap(s2.vf[3], s3.vf[2]);
+	std::swap(s1.vf[3], s3.vf[1]);
 }
 
 } // namespace Effekseer
 
-#endif
+//#endif
 
 #endif // __EFFEKSEER_SIMD4F_GEN_H__