- Turbulence optimized by SIMD
- Add SIMDType.h
- Add SIMD4i

##  Performance improvement
Core i7 1065G7 (x86-64)
- [SSE2] 120ms -> 80ms
- [AVX1] 120ms -> 40ms

Apple A9 1.85 GHz ARM64
- [NEON64] 400ms -> 35ms